Move compareNotNullSigned() and compareNotNullUnsigned() to Bits to utilize vectorized implementation on Java 9+
